プレイを再現したテストケースは別ファイルへ分ける
